+++ src/clj/com/agilecreativity/gh_utils/main.clj
@@ -11,9 +11,11 @@
   (:gen-class))
 
 (defn- load-config
-  "Load config file from the config.edn file"
+  "Load config file that contains username/password information"
   [filename]
-  (edn/read-string (slurp (fs/normalized filename))))
+  (edn/read-string (slurp (-> filename
+                              fs/expand-home
+                              fs/normalized))))
 
 (defn- default-options [& options]
   "Define the sensible default options when creating new Github repository"
@@ -38,7 +40,8 @@
               username (:username config)
               password (:password config)
               auth (str username ":" password)]
-          (let [homepage (str "https://github.com/" (clojure.string/join "/" (list username reponame)))
+          (let [github-prefix "https://github.com/"
+                homepage (str github-prefix (clojure.string/join "/" (list username reponame)))
                 result (t-repos/create-repo reponame
                                             (default-options {:auth auth
                                                               :description (str reponame " by " username)
@@ -46,7 +49,15 @@
             ;; Give user feedback they needed
             (if (:status result)
               (println "Problem creating new repository, errors : " (get-in result [:body :errors]))
-              (println "You have succesfully created new repository at : " (:html_url result))))))
+              (do
+                (let [url (:html_url result)
+                      origin-prefix "git remote add origin "
+                      https-url (str origin-prefix url ".git")
+                      ssh-url (-> https-url
+                                  (clojure.string/replace-first #"https://github.com/" "git@github.com:"))]
+                  (println "You have succesfully created new repository at : " url)
+                  (println "You can track this repository with (https) : " https-url)
+                  (println "You can track this repository with (ssh)   : " ssh-url)))))))
 
       ;; Handle any problem/exception that we may have
       (catch Exception e