+++ src/clj/com/agilecreativity/gh_utils/option.clj
@@ -2,19 +2,21 @@
   (:require [clojure.string :as string])
   (:gen-class))
 
+(def default-config "~/Dropbox/github.edn")
+
 (def options
-  [["-c" "--config CONFIG" :default "config.edn"]
+  [["-c" "--config CONFIG" :default default-config]
    ["-r" "--repo REPO"]
    ["-h" "--help"]])
 
 (defn usage [options-summary]
   (->> ["Create new Github's project from a command line"
         ""
-        "Usage: github-new-repo [options]"
+        "Usage: gh-utils [options]"
         options-summary
         "Options:"
         ""
-        "--config CONFIG full path to the config file e.g. config.edn"
+        "--config CONFIG full path to the config file e.g. ~/Dropbox/github.edn"
         "--repo   REPO   name of repository to be created"
         ""]
        (string/join \newline)))